refactor: move window functions to gui/window.ts
Move changeFullscreen and maximizeWindow to a dedicated window module,
removing scattered appWindow references from util.ts, bootstrap.ts, and globalApi.svelte.ts.
